requirements.txt
feedparser
spacy
nltk
markdown
requests
scikit-learn
streamlit
pillow